server/requirements.txt
cohere==2.6.1
Flask==2.2.2
Flask_Cors==3.0.10
numpy==1.23.3
requests==2.27.1
scikit_learn==1.1.2
twilio==7.14.2
python-dotenv==0.21.0
